* Fixed all bugs in WrappedPacket * Fixed TinyProtocol bugs, cleaned up all the packet managers. * ClientVersion resolvable without ANY dependencies :) We now use TinyProtocol fork of DeprecatedLuke, but I heavily modified it to fit for PacketEvents. * WrappedPacketOutExplosion NPE fixed * ClientVersion.UNRESOLVED constant created, UNKNOWN and UNKNOWN_BUNGEE_SERVER removed. If we can't resolve it, we just return UNRESOLVED * PacketEvents auto resolve client version setting removed, it would default to the server version if one cannot resolve the client version, to avoid any falses we will return UNRESOLVED enum constant if we cannot resolve * We prioritize accessing the protocol version from other dependencies rather than caching it ourselves. If none are available, we will cache it ourselves.
