--- src/main/java/io/github/retrooper/packetevents/nettyhandler/tinyprotocol/TinyProtocol_7.java
@@ -1,906 +1,0 @@
-/*
- *   GNU GENERAL PUBLIC LICENSE
- *      Version 2, June 1991
- *
- *  Copyright (C) 1989, 1991 Free Software Foundation, Inc.,
- *  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
- *  Everyone is permitted to copy and distribute verbatim copies
- *  of this license document, but changing it is not allowed.
- *
- *                             Preamble
- *
- *   The licenses for most software are designed to take away your
- * freedom to share and change it.  By contrast, the GNU General Public
- * License is intended to guarantee your freedom to share and change free
- * software--to make sure the software is free for all its users.  This
- * General Public License applies to most of the Free Software
- * Foundation's software and to any other program whose authors commit to
- * using it.  (Some other Free Software Foundation software is covered by
- * the GNU Lesser General Public License instead.)  You can apply it to
- * your programs, too.
- *
- *   When we speak of free software, we are referring to freedom, not
- * price.  Our General Public Licenses are designed to make sure that you
- * have the freedom to distribute copies of free software (and charge for
- * this service if you wish), that you receive source code or can get it
- * if you want it, that you can change the software or use pieces of it
- * in new free programs; and that you know you can do these things.
- *
- *   To protect your rights, we need to make restrictions that forbid
- * anyone to deny you these rights or to ask you to surrender the rights.
- * These restrictions translate to certain responsibilities for you if you
- * distribute copies of the software, or if you modify it.
- *
- *   For example, if you distribute copies of such a program, whether
- * gratis or for a fee, you must give the recipients all the rights that
- * you have.  You must make sure that they, too, receive or can get the
- * source code.  And you must show them these terms so they know their
- * rights.
- *
- *   We protect your rights with two steps: (1) copyright the software, and
- * (2) offer you this license which gives you legal permission to copy,
- * distribute and/or modify the software.
- *
- *   Also, for each author's protection and ours, we want to make certain
- * that everyone understands that there is no warranty for this free
- * software.  If the software is modified by someone else and passed on, we
- * want its recipients to know that what they have is not the original, so
- * that any problems introduced by others will not reflect on the original
- * authors' reputations.
- *
- *   Finally, any free program is threatened constantly by software
- * patents.  We wish to avoid the danger that redistributors of a free
- * program will individually obtain patent licenses, in effect making the
- * program proprietary.  To prevent this, we have made it clear that any
- * patent must be licensed for everyone's free use or not licensed at all.
- *
- *   The precise terms and conditions for copying, distribution and
- * modification follow.
- *
- *                     GNU GENERAL PUBLIC LICENSE
- *    TERMS AND CONDITIONS FOR COPYING, DISTRIBUTION AND MODIFICATION
- *
- *   0. This License applies to any program or other work which contains
- * a notice placed by the copyright holder saying it may be distributed
- * under the terms of this General Public License.  The "Program", below,
- * refers to any such program or work, and a "work based on the Program"
- * means either the Program or any derivative work under copyright law:
- * that is to say, a work containing the Program or a portion of it,
- * either verbatim or with modifications and/or translated into another
- * language.  (Hereinafter, translation is included without limitation in
- * the term "modification".)  Each licensee is addressed as "you".
- *
- * Activities other than copying, distribution and modification are not
- * covered by this License; they are outside its scope.  The act of
- * running the Program is not restricted, and the output from the Program
- * is covered only if its contents constitute a work based on the
- * Program (independent of having been made by running the Program).
- * Whether that is true depends on what the Program does.
- *
- *   1. You may copy and distribute verbatim copies of the Program's
- * source code as you receive it, in any medium, provided that you
- * conspicuously and appropriately publish on each copy an appropriate
- * copyright notice and disclaimer of warranty; keep intact all the
- * notices that refer to this License and to the absence of any warranty;
- * and give any other recipients of the Program a copy of this License
- * along with the Program.
- *
- * You may charge a fee for the physical act of transferring a copy, and
- * you may at your option offer warranty protection in exchange for a fee.
- *
- *   2. You may modify your copy or copies of the Program or any portion
- * of it, thus forming a work based on the Program, and copy and
- * distribute such modifications or work under the terms of Section 1
- * above, provided that you also meet all of these conditions:
- *
- *     a) You must cause the modified files to carry prominent notices
- *     stating that you changed the files and the date of any change.
- *
- *     b) You must cause any work that you distribute or publish, that in
- *     whole or in part contains or is derived from the Program or any
- *     part thereof, to be licensed as a whole at no charge to all third
- *     parties under the terms of this License.
- *
- *     c) If the modified program normally reads commands interactively
- *     when run, you must cause it, when started running for such
- *     interactive use in the most ordinary way, to print or display an
- *     announcement including an appropriate copyright notice and a
- *     notice that there is no warranty (or else, saying that you provide
- *     a warranty) and that users may redistribute the program under
- *     these conditions, and telling the user how to view a copy of this
- *     License.  (Exception: if the Program itself is interactive but
- *     does not normally print such an announcement, your work based on
- *     the Program is not required to print an announcement.)
- *
- * These requirements apply to the modified work as a whole.  If
- * identifiable sections of that work are not derived from the Program,
- * and can be reasonably considered independent and separate works in
- * themselves, then this License, and its terms, do not apply to those
- * sections when you distribute them as separate works.  But when you
- * distribute the same sections as part of a whole which is a work based
- * on the Program, the distribution of the whole must be on the terms of
- * this License, whose permissions for other licensees extend to the
- * entire whole, and thus to each and every part regardless of who wrote it.
- *
- * Thus, it is not the intent of this section to claim rights or contest
- * your rights to work written entirely by you; rather, the intent is to
- * exercise the right to control the distribution of derivative or
- * collective works based on the Program.
- *
- * In addition, mere aggregation of another work not based on the Program
- * with the Program (or with a work based on the Program) on a volume of
- * a storage or distribution medium does not bring the other work under
- * the scope of this License.
- *
- *   3. You may copy and distribute the Program (or a work based on it,
- * under Section 2) in object code or executable form under the terms of
- * Sections 1 and 2 above provided that you also do one of the following:
- *
- *     a) Accompany it with the complete corresponding machine-readable
- *     source code, which must be distributed under the terms of Sections
- *     1 and 2 above on a medium customarily used for software interchange; or,
- *
- *     b) Accompany it with a written offer, valid for at least three
- *     years, to give any third party, for a charge no more than your
- *     cost of physically performing source distribution, a complete
- *     machine-readable copy of the corresponding source code, to be
- *     distributed under the terms of Sections 1 and 2 above on a medium
- *     customarily used for software interchange; or,
- *
- *     c) Accompany it with the information you received as to the offer
- *     to distribute corresponding source code.  (This alternative is
- *     allowed only for noncommercial distribution and only if you
- *     received the program in object code or executable form with such
- *     an offer, in accord with Subsection b above.)
- *
- * The source code for a work means the preferred form of the work for
- * making modifications to it.  For an executable work, complete source
- * code means all the source code for all modules it contains, plus any
- * associated interface definition files, plus the scripts used to
- * control compilation and installation of the executable.  However, as a
- * special exception, the source code distributed need not include
- * anything that is normally distributed (in either source or binary
- * form) with the major components (compiler, kernel, and so on) of the
- * operating system on which the executable runs, unless that component
- * itself accompanies the executable.
- *
- * If distribution of executable or object code is made by offering
- * access to copy from a designated place, then offering equivalent
- * access to copy the source code from the same place counts as
- * distribution of the source code, even though third parties are not
- * compelled to copy the source along with the object code.
- *
- *   4. You may not copy, modify, sublicense, or distribute the Program
- * except as expressly provided under this License.  Any attempt
- * otherwise to copy, modify, sublicense or distribute the Program is
- * void, and will automatically terminate your rights under this License.
- * However, parties who have received copies, or rights, from you under
- * this License will not have their licenses terminated so long as such
- * parties remain in full compliance.
- *
- *   5. You are not required to accept this License, since you have not
- * signed it.  However, nothing else grants you permission to modify or
- * distribute the Program or its derivative works.  These actions are
- * prohibited by law if you do not accept this License.  Therefore, by
- * modifying or distributing the Program (or any work based on the
- * Program), you indicate your acceptance of this License to do so, and
- * all its terms and conditions for copying, distributing or modifying
- * the Program or works based on it.
- *
- *   6. Each time you redistribute the Program (or any work based on the
- * Program), the recipient automatically receives a license from the
- * original licensor to copy, distribute or modify the Program subject to
- * these terms and conditions.  You may not impose any further
- * restrictions on the recipients' exercise of the rights granted herein.
- * You are not responsible for enforcing compliance by third parties to
- * this License.
- *
- *   7. If, as a consequence of a court judgment or allegation of patent
- * infringement or for any other reason (not limited to patent issues),
- * conditions are imposed on you (whether by court order, agreement or
- * otherwise) that contradict the conditions of this License, they do not
- * excuse you from the conditions of this License.  If you cannot
- * distribute so as to satisfy simultaneously your obligations under this
- * License and any other pertinent obligations, then as a consequence you
- * may not distribute the Program at all.  For example, if a patent
- * license would not permit royalty-free redistribution of the Program by
- * all those who receive copies directly or indirectly through you, then
- * the only way you could satisfy both it and this License would be to
- * refrain entirely from distribution of the Program.
- *
- * If any portion of this section is held invalid or unenforceable under
- * any particular circumstance, the balance of the section is intended to
- * apply and the section as a whole is intended to apply in other
- * circumstances.
- *
- * It is not the purpose of this section to induce you to infringe any
- * patents or other property right claims or to contest validity of any
- * such claims; this section has the sole purpose of protecting the
- * integrity of the free software distribution system, which is
- * implemented by public license practices.  Many people have made
- * generous contributions to the wide range of software distributed
- * through that system in reliance on consistent application of that
- * system; it is up to the author/donor to decide if he or she is willing
- * to distribute software through any other system and a licensee cannot
- * impose that choice.
- *
- * This section is intended to make thoroughly clear what is believed to
- * be a consequence of the rest of this License.
- *
- *   8. If the distribution and/or use of the Program is restricted in
- * certain countries either by patents or by copyrighted interfaces, the
- * original copyright holder who places the Program under this License
- * may add an explicit geographical distribution limitation excluding
- * those countries, so that distribution is permitted only in or among
- * countries not thus excluded.  In such case, this License incorporates
- * the limitation as if written in the body of this License.
- *
- *   9. The Free Software Foundation may publish revised and/or new versions
- * of the General Public License from time to time.  Such new versions will
- * be similar in spirit to the present version, but may differ in detail to
- * address new problems or concerns.
- *
- * Each version is given a distinguishing version number.  If the Program
- * specifies a version number of this License which applies to it and "any
- * later version", you have the option of following the terms and conditions
- * either of that version or of any later version published by the Free
- * Software Foundation.  If the Program does not specify a version number of
- * this License, you may choose any version ever published by the Free Software
- * Foundation.
- *
- *   10. If you wish to incorporate parts of the Program into other free
- * programs whose distribution conditions are different, write to the author
- * to ask for permission.  For software which is copyrighted by the Free
- * Software Foundation, write to the Free Software Foundation; we sometimes
- * make exceptions for this.  Our decision will be guided by the two goals
- * of preserving the free status of all derivatives of our free software and
- * of promoting the sharing and reuse of software generally.
- *
- *                             NO WARRANTY
- *
- *   11. BECAUSE THE PROGRAM IS LICENSED FREE OF CHARGE, THERE IS NO WARRANTY
- * FOR THE PROGRAM, TO THE EXTENT PERMITTED BY APPLICABLE LAW.  EXCEPT WHEN
- * OTHERWISE STATED IN WRITING THE COPYRIGHT HOLDERS AND/OR OTHER PARTIES
- * PROVIDE THE PROGRAM "AS IS" WITHOUT WARRANTY OF ANY KIND, EITHER EXPRESSED
- * OR IMPLIED, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
- * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.  THE ENTIRE RISK AS
- * TO THE QUALITY AND PERFORMANCE OF THE PROGRAM IS WITH YOU.  SHOULD THE
- * PROGRAM PROVE DEFECTIVE, YOU ASSUME THE COST OF ALL NECESSARY SERVICING,
- * REPAIR OR CORRECTION.
- *
- *   12. IN NO EVENT UNLESS REQUIRED BY APPLICABLE LAW OR AGREED TO IN WRITING
- * WILL ANY COPYRIGHT HOLDER, OR ANY OTHER PARTY WHO MAY MODIFY AND/OR
- * REDISTRIBUTE THE PROGRAM AS PERMITTED ABOVE, BE LIABLE TO YOU FOR DAMAGES,
- * INCLUDING ANY GENERAL, SPECIAL, INCIDENTAL OR CONSEQUENTIAL DAMAGES ARISING
- * OUT OF THE USE OR INABILITY TO USE THE PROGRAM (INCLUDING BUT NOT LIMITED
- * TO LOSS OF DATA OR DATA BEING RENDERED INACCURATE OR LOSSES SUSTAINED BY
- * YOU OR THIRD PARTIES OR A FAILURE OF THE PROGRAM TO OPERATE WITH ANY OTHER
- * PROGRAMS), EVEN IF SUCH HOLDER OR OTHER PARTY HAS BEEN ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGES.
- *
- *                      END OF TERMS AND CONDITIONS
- *
- *             How to Apply These Terms to Your New Programs
- *
- *   If you develop a new program, and you want it to be of the greatest
- * possible use to the public, the best way to achieve this is to make it
- * free software which everyone can redistribute and change under these terms.
- *
- *   To do so, attach the following notices to the program.  It is safest
- * to attach them to the start of each source file to most effectively
- * convey the exclusion of warranty; and each file should have at least
- * the "copyright" line and a pointer to where the full notice is found.
- *
- *     <one line to give the program's name and a brief idea of what it does.>
- *     Copyright (C) <year>  <name of author>
- *
- *     This program is free software; you can redistribute it and/or modify
- *     it under the terms of the GNU General Public License as published by
- *     the Free Software Foundation; either version 2 of the License, or
- *     (at your option) any later version.
- *
- *     This program is distributed in the hope that it will be useful,
- *     but WITHOUT ANY WARRANTY; without even the implied warranty of
- *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *     GNU General Public License for more details.
- *
- *     You should have received a copy of the GNU General Public License along
- *     with this program; if not, write to the Free Software Foundation, Inc.,
- *     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- * Also add information on how to contact you by electronic and paper mail.
- *
- * If the program is interactive, make it output a short notice like this
- * when it starts in an interactive mode:
- *
- *     Gnomovision version 69, Copyright (C) year name of author
- *     Gnomovision comes with ABSOLUTELY NO WARRANTY; for details type `show w'.
- *     This is free software, and you are welcome to redistribute it
- *     under certain conditions; type `show c' for details.
- *
- * The hypothetical commands `show w' and `show c' should show the appropriate
- * parts of the General Public License.  Of course, the commands you use may
- * be called something other than `show w' and `show c'; they could even be
- * mouse-clicks or menu items--whatever suits your program.
- *
- * You should also get your employer (if you work as a programmer) or your
- * school, if any, to sign a "copyright disclaimer" for the program, if
- * necessary.  Here is a sample; alter the names:
- *
- *   Yoyodyne, Inc., hereby disclaims all copyright interest in the program
- *   `Gnomovision' (which makes passes at compilers) written by James Hacker.
- *
- *   <signature of Ty Coon>, 1 April 1989
- *   Ty Coon, President of Vice
- *
- * This General Public License does not permit incorporating your program into
- * proprietary programs.  If your program is a subroutine library, you may
- * consider it more useful to permit linking proprietary applications with the
- * library.  If this is what you want to do, use the GNU Lesser General
- * Public License instead of this License.
- */
-
-package io.github.retrooper.packetevents.nettyhandler.tinyprotocol;
-
-import com.google.common.collect.Lists;
-import com.google.common.collect.MapMaker;
-import com.mojang.authlib.GameProfile;
-import io.github.retrooper.packetevents.nettyhandler.NettyPacketManager;
-import io.github.retrooper.packetevents.nettyhandler.tinyprotocol.utils.TinyProtocolReflection;
-import io.github.retrooper.packetevents.nettyhandler.tinyprotocol.utils.TinyProtocolReflection.FieldAccessor;
-import io.github.retrooper.packetevents.nettyhandler.tinyprotocol.utils.TinyProtocolReflection.MethodInvoker;
-import io.github.retrooper.packetevents.utils.nms.NMSUtils;
-import io.github.retrooper.packetevents.utils.reflection.ClassUtil;
-import net.minecraft.util.io.netty.channel.*;
-import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.EventPriority;
-import org.bukkit.event.HandlerList;
-import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerLoginEvent;
-import org.bukkit.event.server.PluginDisableEvent;
-import org.bukkit.plugin.Plugin;
-import org.bukkit.scheduler.BukkitRunnable;
-
-import java.lang.reflect.InvocationTargetException;
-import java.util.*;
-import java.util.concurrent.Future;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.logging.Level;
-
-/**
- * Represents a very tiny alternative to ProtocolLib.
- * <p>
- * It now supports intercepting packets during login and status ping (such as OUT_SERVER_PING)!
- *
- * @author Kristian
- */
-public abstract class TinyProtocol_7 {
-    private static final AtomicInteger ID = new AtomicInteger(0);
-
-    // Used in order to lookup a channel
-    private static final MethodInvoker getPlayerHandle = TinyProtocolReflection.getMethod("{obc}.entity.CraftPlayer", "getHandle");
-    private static final FieldAccessor<Object> getConnection = TinyProtocolReflection.getField("{nms}.EntityPlayer", "playerConnection", Object.class);
-    private static final FieldAccessor<Object> getManager = TinyProtocolReflection.getField("{nms}.PlayerConnection", "networkManager", Object.class);
-    private static final FieldAccessor<Channel> getChannel = TinyProtocolReflection.getField("{nms}.NetworkManager", Channel.class, 0);
-
-    // Looking up ServerConnection
-    private static final Class<Object> minecraftServerClass = TinyProtocolReflection.getUntypedClass("{nms}.MinecraftServer");
-    private static final Class<Object> serverConnectionClass = TinyProtocolReflection.getUntypedClass("{nms}.ServerConnection");
-    private static final FieldAccessor<Object> getMinecraftServer = TinyProtocolReflection.getField("{obc}.CraftServer", minecraftServerClass, 0);
-    private static final FieldAccessor<Object> getServerConnection = TinyProtocolReflection.getField(minecraftServerClass, serverConnectionClass, 0);
-    private static final MethodInvoker getNetworkMarkers = TinyProtocolReflection.getTypedMethod(serverConnectionClass, null, List.class, serverConnectionClass);
-
-    // Packets we have to intercept
-    private static final Class<?> PACKET_LOGIN_IN_START = TinyProtocolReflection.getMinecraftClass("PacketLoginInStart");
-    private static final FieldAccessor<GameProfile> getGameProfile = TinyProtocolReflection.getField(PACKET_LOGIN_IN_START, GameProfile.class, 0);
-
-    // Speedup channel lookup
-    private Map<String, Channel> channelLookup = new MapMaker().weakValues().makeMap();
-    private Listener listener;
-
-    // Channels that have already been removed
-    private Set<Channel> uninjectedChannels = Collections.newSetFromMap(new MapMaker().weakKeys().<Channel, Boolean>makeMap());
-
-    // List of network markers
-    private List<Object> networkManagers;
-
-    // Injected channel handlers
-    private List<Channel> serverChannels = Lists.newArrayList();
-    private ChannelInboundHandlerAdapter serverChannelHandler;
-    private ChannelInitializer<Channel> beginInitProtocol;
-    private ChannelInitializer<Channel> endInitProtocol;
-
-    // Current handler name
-    private String handlerName;
-
-    protected volatile boolean closed;
-    protected Plugin plugin;
-
-    /**
-     * Construct a new instance of TinyProtocol, and start intercepting packets for all connected clients and future clients.
-     * <p>
-     * You can construct multiple instances per plugin.
-     *
-     * @param plugin - the plugin.
-     */
-    public TinyProtocol_7(final Plugin plugin) {
-        this.plugin = plugin;
-
-        // Compute handler name
-        this.handlerName = getHandlerName();
-
-        // Prepare existing players
-        registerBukkitEvents();
-
-        try {
-            registerChannelHandler();
-            registerPlayers(plugin);
-        } catch (IllegalArgumentException ex) {
-            // Damn you, late bind
-            plugin.getLogger().info("[TinyProtocol] Delaying server channel injection due to late bind.");
-
-            new BukkitRunnable() {
-                @Override
-                public void run() {
-                    registerChannelHandler();
-                    registerPlayers(plugin);
-                    plugin.getLogger().info("[TinyProtocol] Late bind injection successful.");
-                }
-            }.runTask(plugin);
-        }
-    }
-
-    private void createServerChannelHandler() {
-        // Handle connected channels
-        endInitProtocol = new ChannelInitializer<Channel>() {
-
-            @Override
-            protected void initChannel(Channel channel) throws Exception {
-                try {
-                    // This can take a while, so we need to stop the main thread from interfering
-                    synchronized (networkManagers) {
-                        // Stop injecting channels
-                        if (!closed) {
-                            channel.eventLoop().submit(() -> injectChannelInternal(channel));
-                        }
-                    }
-                } catch (Exception e) {
-                    plugin.getLogger().log(Level.SEVERE, "Cannot inject incomming channel " + channel, e);
-                }
-            }
-
-        };
-
-        // This is executed before Minecraft's channel handler
-        beginInitProtocol = new ChannelInitializer<Channel>() {
-
-            @Override
-            protected void initChannel(Channel channel) throws Exception {
-                channel.pipeline().addLast(endInitProtocol);
-            }
-
-        };
-
-        serverChannelHandler = new ChannelInboundHandlerAdapter() {
-
-            @Override
-            public void channelRead(ChannelHandlerContext ctx, Object msg) throws Exception {
-                Channel channel = (Channel) msg;
-
-                // Prepare to initialize ths channel
-                channel.pipeline().addFirst(beginInitProtocol);
-                ctx.fireChannelRead(msg);
-            }
-
-        };
-    }
-
-    /**
-     * Register bukkit events.
-     */
-    private void registerBukkitEvents() {
-        listener = new Listener() {
-
-            @EventHandler(priority = EventPriority.LOWEST)
-            public final void onPlayerLogin(PlayerLoginEvent e) {
-                if (closed)
-                    return;
-
-                Channel channel = getChannel(e.getPlayer());
-
-                // Don't inject players that have been explicitly uninjected
-                if (!uninjectedChannels.contains(channel)) {
-                    injectPlayer(e.getPlayer());
-                }
-            }
-
-            @EventHandler
-            public final void onPluginDisable(PluginDisableEvent e) {
-                if (e.getPlugin().equals(plugin)) {
-                    close();
-                }
-            }
-
-        };
-
-        plugin.getServer().getPluginManager().registerEvents(listener, plugin);
-    }
-
-    @SuppressWarnings("unchecked")
-    private void registerChannelHandler() {
-        Object mcServer = null;
-        try {
-            mcServer = NMSUtils.getMinecraftServerInstance();
-        } catch (InvocationTargetException | IllegalAccessException e) {
-            e.printStackTrace();
-        }
-        Object serverConnection = getServerConnection.get(mcServer);
-        boolean looking = true;
-
-        // We need to synchronize against this list
-        networkManagers = (List<Object>) getNetworkMarkers.invoke(null, serverConnection);
-        createServerChannelHandler();
-
-        // Find the correct list, or implicitly throw an exception
-        for (int i = 0; looking; i++) {
-            List<Object> list = TinyProtocolReflection.getField(serverConnection.getClass(), List.class, i).get(serverConnection);
-
-            for (Object item : list) {
-                if (!(item instanceof ChannelFuture))
-                    break;
-
-                // Channel future that contains the server connection
-                Channel serverChannel = ((ChannelFuture) item).channel();
-
-                serverChannels.add(serverChannel);
-                serverChannel.pipeline().addFirst(serverChannelHandler);
-                looking = false;
-            }
-        }
-    }
-
-    private void unregisterChannelHandler() {
-        if (serverChannelHandler == null)
-            return;
-
-        for (Channel serverChannel : serverChannels) {
-            final ChannelPipeline pipeline = serverChannel.pipeline();
-
-            // Remove channel handler
-            serverChannel.eventLoop().execute(new Runnable() {
-
-                @Override
-                public void run() {
-                    try {
-                        pipeline.remove(serverChannelHandler);
-                    } catch (NoSuchElementException e) {
-                        // That's fine
-                    }
-                }
-
-            });
-        }
-    }
-
-    private void registerPlayers(Plugin plugin) {
-        for (Player player : plugin.getServer().getOnlinePlayers()) {
-            injectPlayer(player);
-        }
-    }
-
-    /**
-     * Invoked when the server is starting to send a packet to a player.
-     * <p>
-     * Note that this is not executed on the main thread.
-     *
-     * @param receiver - the receiving player, NULL for early login/status packets.
-     * @param channel  - the channel that received the packet. Never NULL.
-     * @param packet   - the packet being sent.
-     * @return The packet to send instead, or NULL to cancel the transmission.
-     */
-    public Object onPacketOutAsync(Player receiver, Object channel, Object packet) {
-        return packet;
-    }
-
-    /**
-     * Invoked when the server has received a packet from a given player.
-     * <p>
-     * Use {@link Channel#remoteAddress()} to get the remote address of the client.
-     *
-     * @param sender  - the player that sent the packet, NULL for early login/status packets.
-     * @param channel - channel that received the packet. Never NULL.
-     * @param packet  - the packet being received.
-     * @return The packet to recieve instead, or NULL to cancel.
-     */
-    public Object onPacketInAsync(Player sender, Object channel, Object packet) {
-        return packet;
-    }
-
-    /**
-     * Send a packet to a particular player.
-     * <p>
-     * Note that {@link #onPacketOutAsync(Player, Object, Object)} will be invoked with this packet.
-     *
-     * @param player - the destination player.
-     * @param packet - the packet to send.
-     */
-    public void sendPacket(Player player, Object packet) {
-        sendPacket(getChannel(player), packet);
-    }
-
-    /**
-     * Send a packet to a particular client.
-     * <p>
-     * Note that {@link #onPacketOutAsync(Player, Object, Object)} will be invoked with this packet.
-     *
-     * @param channel - client identified by a channel.
-     * @param packet  - the packet to send.
-     */
-    public void sendPacket(Object channel, Object packet) {
-        Channel nettyChannel = (Channel)channel;
-        nettyChannel.pipeline().writeAndFlush(packet);
-    }
-
-    /**
-     * Pretend that a given packet has been received from a player.
-     * <p>
-     * Note that {@link #onPacketInAsync(Player, Object, Object)} will be invoked with this packet.
-     *
-     * @param player - the player that sent the packet.
-     * @param packet - the packet that will be received by the server.
-     */
-    public void receivePacket(Player player, Object packet) {
-        receivePacket(getChannel(player), packet);
-    }
-
-    /**
-     * Pretend that a given packet has been received from a given client.
-     * <p>
-     * Note that {@link #onPacketInAsync(Player, Object, Object)} will be invoked with this packet.
-     *
-     * @param channel - client identified by a channel.
-     * @param packet  - the packet that will be received by the server.
-     */
-    public void receivePacket(Channel channel, Object packet) {
-        channel.pipeline().context("encoder").fireChannelRead(packet);
-    }
-
-    /**
-     * Retrieve the name of the channel injector, default implementation is "tiny-" + plugin name + "-" + a unique ID.
-     * <p>
-     * Note that this method will only be invoked once. It is no longer necessary to override this to support multiple instances.
-     *
-     * @return A unique channel handler name.
-     */
-    protected String getHandlerName() {
-        return "tiny-" + plugin.getName() + "-" + ID.incrementAndGet();
-    }
-
-    /**
-     * Add a custom channel handler to the given player's channel pipeline, allowing us to intercept sent and received packets.
-     * <p>
-     * This will automatically be called when a player has logged in.
-     *
-     * @param player - the player to inject.
-     */
-    public void injectPlayer(Player player) {
-        injectChannelInternal(getChannel(player)).player = player;
-    }
-
-    public void injectPlayerAsync(Player player) {
-        injectChannelInternalAsync(getChannel(player)).player = player;
-    }
-
-    /**
-     * Add a custom channel handler to the given channel.
-     *
-     * @param channel - the channel to inject.
-     * @return The intercepted channel, or NULL if it has already been injected.
-     */
-    public void injectChannel(Channel channel) {
-        injectChannelInternal(channel);
-    }
-
-    public void injectChannelAsync(Channel channel) {
-        injectChannelInternalAsync(channel);
-    }
-
-    /**
-     * Add a custom channel handler to the given channel.
-     *
-     * @param channel - the channel to inject.
-     * @return The packet interceptor.
-     */
-    private PacketInterceptor injectChannelInternal(Channel channel) {
-        try {
-            PacketInterceptor interceptor = (PacketInterceptor) channel.pipeline().get(handlerName);
-
-            // Inject our packet interceptor
-            if (interceptor == null) {
-                interceptor = new PacketInterceptor();
-                channel.pipeline().addBefore("packet_handler", handlerName, interceptor);
-                uninjectedChannels.remove(channel);
-            }
-
-            return interceptor;
-        } catch (IllegalArgumentException e) {
-            // Try again
-            return (PacketInterceptor) channel.pipeline().get(handlerName);
-        }
-    }
-
-    private PacketInterceptor injectChannelInternalAsync(Channel channel) {
-        try {
-            PacketInterceptor interceptor = (PacketInterceptor) channel.pipeline().get(handlerName);
-
-            // Inject our packet interceptor
-            if (interceptor == null) {
-                final PacketInterceptor newInterceptor = new PacketInterceptor();
-                interceptor = newInterceptor;
-                uninjectedChannels.remove(channel);
-                NettyPacketManager.executorService.submit(new Runnable() {
-                    @Override
-                    public void run() {
-                        channel.pipeline().addBefore("packet_handler", handlerName, newInterceptor);
-                    }
-                });
-            }
-
-            return interceptor;
-        } catch (IllegalArgumentException e) {
-            // Try again
-            return (PacketInterceptor) channel.pipeline().get(handlerName);
-        }
-    }
-
-    /**
-     * Retrieve the Netty channel associated with a player. This is cached.
-     *
-     * @param player - the player.
-     * @return The Netty channel.
-     */
-    public Channel getChannel(Player player) {
-        Channel channel = channelLookup.get(player.getName());
-
-        // Lookup channel again
-        if (channel == null) {
-            Object connection = getConnection.get(getPlayerHandle.invoke(player));
-            Object manager = getManager.get(connection);
-
-            channelLookup.put(player.getName(), channel = getChannel.get(manager));
-        }
-
-        return channel;
-    }
-
-    /**
-     * Uninject a specific player.
-     *
-     * @param player - the injected player.
-     */
-    public void uninjectPlayer(Player player) {
-        uninjectChannel(getChannel(player));
-    }
-
-    public Future<?> uninjectPlayerAsync(Player player) {
-        return uninjectChannelAsync(getChannel(player));
-    }
-
-    /**
-     * Uninject a specific channel.
-     * <p>
-     * This will also disable the automatic channel injection that occurs when a player has properly logged in.
-     *
-     * @param channel - the injected channel.
-     */
-    public void uninjectChannel(final Channel channel) {
-        // No need to guard against this if we're closing
-        if (!closed) {
-            uninjectedChannels.add(channel);
-        }
-
-        // See ChannelInjector in ProtocolLib, line 590
-        channel.pipeline().remove(handlerName);
-    }
-
-    public Future<?> uninjectChannelAsync(Channel channel) {
-        if (!closed) {
-            uninjectedChannels.add(channel);
-        }
-
-        return NettyPacketManager.executorService.submit(new Runnable() {
-            @Override
-            public void run() {
-                channel.pipeline().remove(handlerName);
-            }
-        });
-    }
-
-    /**
-     * Determine if the given player has been injected by TinyProtocol.
-     *
-     * @param player - the player.
-     * @return TRUE if it is, FALSE otherwise.
-     */
-    public boolean hasInjected(Player player) {
-        return hasInjected(getChannel(player));
-    }
-
-    /**
-     * Determine if the given channel has been injected by TinyProtocol.
-     *
-     * @param channel - the channel.
-     * @return TRUE if it is, FALSE otherwise.
-     */
-    public boolean hasInjected(Channel channel) {
-        return channel.pipeline().get(handlerName) != null;
-    }
-
-    /**
-     * Cease listening for packets. This is called automatically when your plugin is disabled.
-     */
-    public final void close() {
-        if (!closed) {
-            closed = true;
-
-            // Remove our handlers
-            for (Player player : plugin.getServer().getOnlinePlayers()) {
-                uninjectPlayer(player);
-            }
-
-            // Clean up Bukkit
-            HandlerList.unregisterAll(listener);
-            unregisterChannelHandler();
-        }
-    }
-
-    /**
-     * Channel handler that is inserted into the player's channel pipeline, allowing us to intercept sent and received packets.
-     *
-     * @author Kristian
-     */
-    private final class PacketInterceptor extends ChannelDuplexHandler {
-        // Updated by the login event
-        public volatile Player player;
-
-        @Override
-        public void channelRead(ChannelHandlerContext ctx, Object msg) throws Exception {
-            // Intercept channel
-            final Channel channel = ctx.channel();
-            handleLoginStart(channel, msg);
-            try {
-                msg = onPacketInAsync(player, channel, msg);
-            } catch (Exception e) {
-                plugin.getLogger().log(Level.SEVERE, "Error in onPacketInAsync().", e);
-            }
-
-            if (msg != null) {
-                super.channelRead(ctx, msg);
-                NettyPacketManager.postRead(player, msg);
-            }
-
-        }
-
-        @Override
-        public void write(ChannelHandlerContext ctx, Object msg, ChannelPromise promise) throws Exception {
-            try {
-                msg = onPacketOutAsync(player, ctx.channel(), msg);
-            } catch (Exception e) {
-                plugin.getLogger().log(Level.SEVERE, "Error in onPacketOutAsync().", e);
-            }
-
-            if (msg != null) {
-                super.write(ctx, msg, promise);
-                NettyPacketManager.postSend(player, msg);
-            }
-        }
-
-        private void handleLoginStart(Channel channel, Object packet) {
-            if (PACKET_LOGIN_IN_START.isInstance(packet)) {
-                GameProfile profile = getGameProfile.get(packet);
-                channelLookup.put(profile.getName(), channel);
-            }
-        }
-    }
-}
